actually be greedy (#123)
* rollups take a single item at a time
* greatly improve memory allocations and GC pressure

+++ lrdb/metric_pack_estimate.sql.go
@@ -58,7 +58,7 @@ DO UPDATE SET
 
 type UpsertMetricPackEstimateParams struct {
 	OrganizationID uuid.UUID `json:"organization_id"`
-	FrequencyMs    int64     `json:"frequency_ms"`
+	FrequencyMs    int32     `json:"frequency_ms"`
 	TargetRecords  *int64    `json:"target_records"`
 }
